Refactored changeAnnotations function. Changed annotation level to be integer rather than double.  #refactor #story[670]

+++ src/main/java/seng302/Controller.java
@@ -84,7 +84,7 @@ public class Controller implements Initializable {
         annotationsSlider.valueProperty().addListener(new ChangeListener<Number>() {
             @Override
             public void changed(ObservableValue<? extends Number> observable, Number oldValue, Number newValue) {
-                display.changeAnnotations((double) newValue);
+                display.changeAnnotations(newValue.intValue());
             }
         });
         annotationsSlider.adjustValue(annotationsSlider.getMax());
